feat: add support of default directory
The default directory is use when the user initialized for the first
time a file or directory property.
Without the default directory pattern, the selection dialog display the
user home directory at each initialization. Thus, if all the files to
initialize are in the same directory, the user shall se the same path
from its user home at each time.
With the default directory pattern, the selection dialog display the
user home directory only at the first initializations. For the other
initializations, the previous directory is displayed by default.